requirements.txt
python-dotenv>=1.0.1
dashscope>=1.20.0
